code-runner/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
docker==7.0.0
celery==5.3.6
redis==5.0.1
boto3==1.34.57
python-dotenv==1.0.1